refactor(client): extract collection operation boundary
